refactoring loadConfig
