+++ packages/service/test/config.test.js
@@ -2,7 +2,7 @@
 
 require('./helper')
 const { test } = require('tap')
-const { buildServer } = require('..')
+const { buildServer, platformaticService } = require('..')
 const { loadConfig } = require('../lib/load-config')
 const { request } = require('undici')
 const { join } = require('path')
@@ -297,28 +297,17 @@ test('config reloads', async ({ teardown, equal, pass, same }) => {
 })
 
 test('can merge provided config with default config', async ({ plan, same }) => {
-  plan(6)
+  plan(3)
   const configFile = join(__dirname, 'fixtures', 'custom-port-placeholder.json')
   const defaultConfig = {
-    mergeDefaults: true,
     onMissingEnv (key) {
       same(key, 'A_CUSTOM_PORT')
       return '42'
     }
   }
 
-  {
-    const config = await loadConfig({}, ['-c', configFile], defaultConfig)
-    same(config.configManager.current.server.port, 42)
-    // This comes from the default config.
-    same(config.configManager.schemaOptions.useDefaults, true)
-  }
-
-  {
-    defaultConfig.mergeDefaults = false
-    const config = await loadConfig({}, ['-c', configFile], defaultConfig)
-    same(config.configManager.current.server.port, 42)
-    // This comes from the default config.
-    same(config.configManager.schemaOptions.useDefaults, undefined)
-  }
+  const config = await loadConfig({}, ['-c', configFile], platformaticService, 'service', defaultConfig)
+  same(config.configManager.current.server.port, 42)
+  // This comes from the default config.
+  same(config.configManager.schemaOptions.useDefaults, true)
 })
